[Xcode 4.6] Minor warning cleanup and project tweaks.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5630 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/OOStringExpander.m
@@ -1079,7 +1079,7 @@ static NSString *NewRandomDigrams(void)
 	unsigned length = (gen_rnd_number() % 4) + 1;
 	if ((gen_rnd_number() % 5) < ((length == 1) ? 3 : 1))  ++length;	// Make two-letter names rarer and 10-letter names happen sometimes
 	NSString *digrams = [[UNIVERSE descriptions] objectForKey:@"digrams"];
-	unsigned count = [digrams length] / 2;
+	NSUInteger count = [digrams length] / 2;
 	NSMutableString *name = [NSMutableString stringWithCapacity:length * 2];
 	
 	for (unsigned i = 0; i != length; ++i)
